backend/requirements.txt
# requirements

fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
Pillow
